ld: revert two upstream commits related to interpretation of ".".
This change reverts commits d2667025dd30611514810c28bee9709e4623012a and 975f8a9e3144e4d3d3f391e907c8bf94b23dc8b6
  - https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=d2667025dd30611514810c28bee9709e4623012a
  - https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=975f8a9e3144e4d3d3f391e907c8bf94b23dc8b6
from binutils-2_27-branch.

Commit d2667025dd30611514810c28bee9709e4623012a (dated 2012-06-15)
changed the interpretation of "." when it appears in an assignment
statement outside of an output section statement. This change
interacts with the linux kernel 3.8.* builds for amd64 and results
in an unbootable kernel. This issue was introduced in binutils 2.23,
and we have locally reverted this change on every upgrade since then.

Commit 975f8a9e3144e4d3d3f391e907c8bf94b23dc8b6 (dated 2015-09-18)
is a followup to d2667025dd30611514810c28bee9709e4623012a. It fixes
pr/18963, which was caused by the first commit, by delaying some of
the symbol value computations.
  - https://sourceware.org/bugzilla/show_bug.cgi?id=18963

Change-Id: I1eb6177d79206c8b321ccf0ee0c1e293fc94093b