Make sure errata fix is done, whether or not flags are passed.
Currently the reloc stubs errata code will only be run if a
fix-errata flag is passed to the linker.  This CL makes sure the
errata fix is applied, whether or not the linker flag is passed (this
will prevent a runtime bug).

Bug: https://b/63044552
Test: Built binutils for linux for all 4 architectures; built Angler
image & kernel with updated binutils and tested them successfully on
an Angler device.

Change-Id: Ieafdc8a12ae78eee14efc5db3c6a64468fec125c