gold: fix PR gold/21491 - errata workaround can produce broken images.
The problem is caused by the fact that gold is relocating the stubs
for an entire output section when it processes the relocations for a
particular input section that happened to be designated as the stub
table "owner". The Relocate_task for that input section may or may not
run before the Relocate_task for another input section that contains
the code that needs the erratum fix, but doesn't "own" the stub
table. If it runs before (or might even race with) that other task, it
ends up with a copy of the unrelocated original instruction.

In other words - when calling fix_errata() from
do_relocate_sections(), gold is going through the list of errata stubs
that are associated only with that object. This routine updates the
stored original instruction and replaces it in the output view with a
branch to the stub. Later, as gold is going through the object file's
input sections, it then checks for stub tables "owned" by each input
section, and writes out all the stubs from that stub table, regardless
of what object file each stub is associated with.

Fixed by relocating the erratum stub only after the corresponding
errata spot is fixed. That is to have fix_errata() call
Stub_table::relocate_erratum_stub() for each stub.

        PR gold/21491
        * aarch64.cc (Erratum_stub::invalidate_erratum_stub): New method.
        (Erratum_stub::is_invalidated_erratum_stub): New method.
        (Stub_table::relocate_reloc_stub): Renamed from "relocate_stub".
        (Stub_table::relocate_reloc_stubs): Renamed from "relocate_stubs".
        (Stub_table::relocate_erratum_stub): New method.
        (AArch64_relobj::fix_errata_and_relocate_erratum_stubs): Renamed from
        "fix_errata".
        (Target_aarch64::relocate_reloc_stub): Renamed from "relocate_stub".

This change is backport of commit df2f63a6a0fc19c54e58aab8afe262baf3cb1a3c
  - https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=df2f63a6a0fc19c54e58aab8afe262baf3cb1a3c
to binutils-2_27-branch.

Change-Id: I1f70f82f0330b4249589d05a86d5d9daac6e9b93