Add a script for building just binutils.
In contrast to building binutils via the GCC build, each target here
takes only 1 minute instead of 10. Given that we've stopped
supporting GCC, it's also nice to not have to rebuild something that
shouldn't have changed anyway.

This also has the advantage of building to a specific location in the
out directory rather than /tmp, meaning it's easy to cd into the
build directory and just run make rather than rerunning the full
build script.

Test: ./build.py for all architectures, linux, win, win64
Bug: None
Change-Id: I067d84c5439c56004fc12de496f8f66fba7f7bf7