Fix RELRO calculation bug when .data segment with RELRO was not last
Gold doesn't know how to sort unique section -> segment mapped segments
correctly relative to normal segments with equivalent flags. Because of this,
adding a new data segment may result in the segment containing RELRO sections
not being last and the GNU_RELRO size not being computed correctly. This fix
ensures that we do not recompute increase_relro when setting section addresses
for segments that do not contain any relro sections.

Change-Id: Ic401270df53ba2155e791b31cf9b16420a88464a