bfd: PR22829, objcopy/strip removes PT_GNU_RELRO from lld binaries
lld lays out the relro segment differently to GNU ld, not bothering to
include the first few bytes of .got.plt and padding out to a page at
the end of the segment.  This patch teaches binutils to recognize the
different (and somewhat inferior) layout as valid.

    bfd/
        PR 22829
        * elf.c (assign_file_positions_for_non_load_sections): Rewrite
        PT_GNU_RELRO setup.
    ld/
        * testsuite/ld-x86-64/pr14207.d: Adjust relro p_filesz.

This change is backport of commit f2731e0c374e5323ce4cdae2bcc7b7fe22da1a6f
  - https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=f2731e0c374e5323ce4cdae2bcc7b7fe22da1a6f
to binutils-2.27.

Change-Id: I9d598adb1a0ef4778f0f29d0faf19c7c8112a74a