[GOLD] Fix _GLOBAL_OFFSET_TABLE_ value for large GOTs
Gold resolves GOT-relative relocs relative to the GOT base + 0x8000 when the GOT
is larger than 0x8000. However, previously the _GLOBAL_OFFSET_TABLE_ symbol was
set to GOT base + 0x8000 when the .got.plt was larger than 0x8000. This patch
makes both checks use the size of the .got section so that they agree when to
add 0x8000.

Change-Id: I7df95dc390b47a91f6250348bfa357f31f5dbfc2